# src/api.yaml
openapi: 3.0.3
info:
  title: HD1 (Holodeck One) API
  description: |
    Revolutionary API-first 3D coordinate system with universal world boundaries.
    Single source of truth for sessions, worlds, objects, and real-time control.
    
    **CORE PRINCIPLES:**
    - Everything API-driven (zero shell commands)
    - Universal coordinate system [-12, +12] on all axes
    - Objects honor world boundaries absolutely
    - Named object persistence with full lifecycle
    - Real-time WebSocket + REST hybrid
    
    **ARCHITECTURE:**
    - Specification drives code generation
    - Build fails if handlers missing
    - Zero manual routing - auto-generated from spec
    
  version: 1.0.0
  contact:
    name: Holodeck One 3D Engine
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Get session details
      description: Retrieve detailed session information including all objects
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/sessions/get.go"
      x-function: "GetSession"
      
    delete:
      operationId: deleteSession
      tags: [Sessions]
      summary: Terminate session
      description: Permanently delete session and all associated objects
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/sessions/delete.go"
      x-function: "DeleteSession"

  # ========================================
  # SESSION GRAPH SESSION GRAPH & ROOM MANAGEMENT CHANNEL MANAGEMENT
  # TCP-level reliability with channel-based architecture
  # ========================================
...
  /scenes/{sceneId}:
    post:
      operationId: loadScene
      tags: [Scenes]
      summary: Load scene into session
      description: |
        Loads a predefined scene (HD1 script) into the specified session.
        Clears existing objects and applies scene configuration.
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: string
          description: Scene identifier (empty, anime-ui, ultimate, basic-shapes)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  description: Target session for scene loading
              required:
                - session_id
      responses:
        '200':
          description: Scene loaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scene_id:
                    type: string
                  session_id:
                    type: string
                  objects_created:
                    type: integer
                  message:
                    type: string
        '404':
          description: Scene not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid session or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-handler: "api/scenes/load.go"
      x-function: "LoadScene"

  # ========================================
  # SCENE FORKING & MANAGEMENT
  # ========================================
  /sessions/{sessionId}/scenes/save:
    post:
      operationId: saveSceneFromSession
      tags: [Scenes]
      summary: Save session state as new scene
      description: |
        Creates a new scene script from current session objects.
        Generates a .sh script in share/glibsh/scenes/ with all session objects.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scene_id:
                  type: string
                  description: ID for new scene (becomes filename)
                  example: "my-custom-scene"
                name:
                  type: string
                  description: Human-readable scene name
                  example: "My Custom Scene"
                description:
                  type: string
                  description: Scene description
                  example: "Custom scene with modified objects"
                overwrite:
                  type: boolean
                  description: Allow overwriting existing scene
                  default: false
              required: [scene_id, name, description]
      responses:
        '201':
          description: Scene saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  scene_id:
                    type: string
                  scene_path:
                    type: string
                  objects_exported:
                    type: integer
        '400':
          description: Invalid request or scene already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Scene already exists (use overwrite=true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-handler: "api/scenes/save.go"
      x-function: "SaveSceneFromSession"
...
  /sessions/{sessionId}/scene/state/save:
    post:
      operationId: saveSceneState
      tags: [SceneGraph]
      summary: Save scene state snapshot
      description: |
        Creates a named snapshot of the current scene state for later restoration.
        Includes entity hierarchy, scene settings, and asset references.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                snapshot_name:
                  type: string
                  description: Name for the scene state snapshot
                description:
                  type: string
                  description: Optional description of the snapshot
                include_assets:
                  type: boolean
                  description: Include asset references in snapshot
                  default: true
              required: [snapshot_name]
      responses:
        '200':
          description: Scene state snapshot saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  snapshot_id:
                    type: string
                  entities_saved:
                    type: integer
                  timestamp:
                    type: string
        '400':
          description: Invalid snapshot parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/SessionNotFound'
      x-handler: "api/scenegraph/state.go"
      x-function: "SaveSceneState"
...
  /sessions/{sessionId}/scenes:
    get:
      operationId: listSessionScenes
      tags: [SceneGraph]
      summary: List all scenes in session
      description: |
        Returns all scene instances within a session for multi-scene
        composition scenarios.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session scenes listed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scenes:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionScene'
                  active_scene:
                    type: string
                    description: ID of currently active scene
        '404':
          $ref: '#/components/responses/SessionNotFound'
      x-handler: "api/scenegraph/scenes.go"
      x-function: "ListSessionScenes"

    post:
      operationId: createSessionScene
      tags: [SceneGraph]
      summary: Create new scene in session
      description: |
        Creates a new scene instance within the session for multi-scene
        composition workflows.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scene_name:
                  type: string
                  description: Name for the new scene
                template:
                  type: string
                  description: Scene template to use
                  enum: [empty, basic, lighting, physics]
                  default: empty
                set_active:
                  type: boolean
                  description: Set as active scene after creation
                  default: true
              required: [scene_name]
      responses:
        '201':
          description: Scene created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scene_id:
                    type: string
                  scene_name:
                    type: string
                  is_active:
                    type: boolean
        '400':
          description: Invalid scene parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/SessionNotFound'
      x-handler: "api/scenegraph/scenes.go"
      x-function: "CreateSessionScene"
...
components:
  schemas:
...
    SessionSummary:
      type: object
      required: [id, created, world_initialized, object_count, status]
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        world_initialized:
          type: boolean
        object_count:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [active, inactive, error]
    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/SessionSummary'
        - type: object
          properties:
            objects:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ObjectDetail'

    # Environment Schemas
...
    SessionScene:
      type: object
      required: [sceneId, sceneName, isActive, entities]
      properties:
        sceneId:
          type: string
          description: "Unique scene identifier"
          example: "scene_abc123"
        sceneName:
          type: string
          description: "Scene display name"
          example: "Main Game Scene"
        isActive:
          type: boolean
          description: "Whether this scene is currently active"
          example: true
        entities:
          type: integer
          minimum: 0
          description: "Number of entities in this scene"
          example: 15
        template:
          type: string
          description: "Scene template used for creation"
          example: "basic"
        createdAt:
          type: string
          format: date-time
          description: "Scene creation timestamp"
          example: "2025-07-01T17:15:30.123Z"
        lastModified:
          type: string
          format: date-time
          description: "Last modification timestamp"
          example: "2025-07-01T17:20:45.456Z"
...
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    SessionNotFound:
      description: Session not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ========================================
# EXTENSION PROPERTIES
# ========================================
...